put types shared by server and application code in types.d.ts
