+++ package.json
@@ -31,7 +31,7 @@
     "test": "yarn node --experimental-vm-modules ./node_modules/jest-cli/bin/jest --testPathIgnorePatterns=black-box --forceExit",
     "test:black-box": "rimraf dist && rimraf out && yarn build && yarn node --experimental-vm-modules ./node_modules/jest-cli/bin/jest black-box --forceExit --coverage=false",
     "test:mutants": "stryker run stryker.config.json",
-    "build": "tsc && copyfiles src/client/** dist && copyfiles templates/** dist",
+    "build": "tsc && copyfiles -f 'src/client/**' dist/client",
     "prepack": "yarn build",
     "release": "npx changeset publish",
     "prepare": "husky install",
