--- src/server/types.ts
@@ -1,7 +1,0 @@
-export interface OpenApiParameters {
-  in: "body" | "cookie" | "formData" | "header" | "path" | "query";
-  name: string;
-  schema?: {
-    type: string;
-  };
-}
